// fhf-datafangst-client/src/generated/openapi/models/update-vessel.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { EngineType } from './engine-type';

/**
 * 
 * @export
 * @interface UpdateVessel
 */
export interface UpdateVessel {
    /**
     * 
     * @type {number}
     * @memberof UpdateVessel
     */
    'enginePower'?: number | null;
    /**
     * 
     * @type {number}
     * @memberof UpdateVessel
     */
    'engineBuildingYear'?: number | null;
    /**
     * 
     * @type {number}
     * @memberof UpdateVessel
     */
    'auxiliaryEnginePower'?: number | null;
    /**
     * 
     * @type {number}
     * @memberof UpdateVessel
     */
    'auxiliaryEngineBuildingYear'?: number | null;
    /**
     * 
     * @type {number}
     * @memberof UpdateVessel
     */
    'boilerEnginePower'?: number | null;
    /**
     * 
     * @type {number}
     * @memberof UpdateVessel
     */
    'boilerEngineBuildingYear'?: number | null;
    /**
     * 
     * @type {EngineType}
     * @memberof UpdateVessel
     */
    'engineType'?: EngineType;
    /**
     * 
     * @type {number}
     * @memberof UpdateVessel
     */
    'engineRpm'?: number | null;
    /**
     * 
     * @type {number}
     * @memberof UpdateVessel
     */
    'degreeOfElectrification'?: number | null;
    /**
     * 
     * @type {number}
     * @memberof UpdateVessel
     */
    'serviceSpeed'?: number | null;
}